chore: refactored app.js
