+++ src/app.js
@@ -1,40 +1,18 @@
 import Fastify from 'fastify';
-import { getDB, saveDB } from './utils/db/index.js';
-import { v4 } from 'uuid';
+import { createBlog } from './services/blogs/create-blog.js';
+import { general } from './services/general/index.js';
 
 const prefix = '/api';
 
 export async function build () {
   // initialize fastify
   const fastify = Fastify({ logger: true });
-  fastify.get(prefix, async (request, reply) => {
-    return { success: true };
-  });
+
+  // returns { success : true }from /general/index.js
+  fastify.get(prefix, general);
 
   // create blog
-  fastify.post(`${prefix}/blog`, async (request, reply) => {
-    const { body } = request;
-    const { title, desc } = body;
-    const db = await getDB();
-
-    const id = v4();
-
-    const blog = {
-      title,
-      desc,
-      createdDate: new Date().getTime(),
-      updatedDate: new Date().getTime()
-    };
-
-    db.blogs[id] = blog;
-
-    await saveDB(db);
-
-    return {
-      id,
-      ...blog
-    };
-  });
+  fastify.post(`${prefix}/blog`, createBlog);
 
   return fastify;
 }
